SearchBox fix
Add mediaPageEnum, enum that add me the abilty to give a category name and get the class that response to this category.
Add formatter to classes.
Add methods to Utils, first method get class name and driver and return instance of the page that response on it.
Add method that get from the url the category of the page.
Add two method that delete unessery information from title of the media.